Configure the script using environment variables
Putting configuration, especially private configuration, into a config
file in a repo isn't the best thing. Instead, we want to be able to
inject this configuration using environment variables. This adds support
for that, documents the environment variables, and renames some of them
to make things clearer.

In addition, it improves the summary in the README to make it more
accurate.