Use a more functional approach
Using globals is… not the cleanest way to do this, and means we're
relying on side-effects for core behaviour. Which, given the original
version was (one assumes) trying to ensure consistency between what's
published to MQTT and Prometheus, sort of makes sense. But it makes the
program very messy and side-effect-y.

This commit makes the main method for querying the datalogger return
the results, renames it (and various other things) to be clearer in
their intent, and tweaks some of the logging.