Simplify retrying calls that fail
We want to be able to easily retry reading modbus registers, but
previously this was cumbersome and obfuscated the program logic with
retry clutter.

This way, we can just say “retry this” and the actual mechanics of how
that happens are dealt with elsewhere.

I'm not wild about just passing a function with no args, but I'm also
not wild about the apparently-more-pythonic “pass the function
identifier and then forward some of the outer function's parameters”
way, either, because what if there's a clash in parameter names or
something?

This also tweaks some logging so that we get things like traces
for exceptions.