Return multi-byte registers' values as one metric
Some registers have a high and low byte which need to be combined into
one output value. This abstracts the actual reading of the registers
into a dedicated method. This method can then decide if it needs to
handle multi-byte registers (one at a time, slow) or not (fast,
batch-able).