Describe registers using a lightweight DSL
Instead of using a plain tuple, we're going to use a small
domain-specific language to describe the structure of the registers
we'd like to query.

By doing this with namedtuples, this is backwards-compatible with the
previous code, but offers the ability to gradually extract
functionality into the DSL objects. This will allow us to write much
cleaner, more functional code with fewer side effects.

I'm not _wild_ about the clutter of `RS`s and `R`s and `FRS`s
throughout the register config, but it's lighter-weight than a
dictionary, more explicit than a tuple, and allows us to later move to
not having to traverse the structure so much as have the structure do
the work for us.