Make RegisterSets responsible for reading data
One of the advantages of using a DSL is that we can make the building
blocks responsible for the execution of the functionality they describe.

This change makes RegisterSets read data from the modbus interface.
Doing this means we can reduce the amount of conditional control flow
(for example, by using composition instead of logic) in the main
program flow, and generally make everything more explicit.