Simplify reading metrics
The consumer of metrics shouldn't need to know that they have to zip
the list of values with a list of register definitions to be able to
know which result is which.

This moves the responsibility for combining values with their
definitions to the RegisterSets, further hiding the act of reading from
the data logger, cleaning up the call-site, and making everything a
little more functional.

Some things of note:

- Because `map()` produces an iterator, and `chain.from_iterable()`
  creates a lazily-evaluated iterable, query_datalogger() now actually
  returns a lazy-loading iterable of results, meaning the data logger
  isn't accessed until the results are consumed by either the
  Prometheus collector or the MQTT publisher. Which is kind of cool? But
  could also be confusing.
- I'm not wild about retrying the _whole_ `read_metrics` call,
  particularly for FormattedRegisterSets (which make multiple
  underlying calls for registers). Managing the retries at the top
  level feels like the best place, though, and there's not a
  particularly clean way to inject the retry handler to the bottom
  level. And retrying inside a list comprehension feels gross.
- Why is itertools.chain.from_iterable() not flatten()? Because Python.