Only open one modbus connection to the data logger
There are a number of cases (particularly with changes I'm in the
process of making) where we can inadvertently create a situation where
we open and try to read from multiple socket connections, which breaks
all of them, and makes this prone to failing with errors about broken
sockets and so on.

Instead, this creates one modbus instance at the start of the script,
and uses it throughout. This means that all socket management is
handled by PySolarmanV5, and our code is both more robust and faster
because we're not constantly opening new connections to the data logger.

Oh, and it yet again improves the separation of concerns, too, by
taking responsibility for creating the modbus connection away from
get_metrics_reader() and hoisting it to the top level of the script.