Simplify register config and reading
We can significantly simplify how we describe and read registers by
using a mix of generators (for reading arbitrary chunks of data) and
`PySolarmanV5._format_response()` for every register, even those we
don't want to format (which is fine, because it does nothing in that
case).

Because of this, we can:

- Use the same, simple `RegisterSet` to describe and read _any_ group
  of `Registers`
- Make all registers “formatted” at no cost, meaning we can have signed
  and multi-byte registers mixed with “normal” ones, as they are in
  reality. And non-formatted registers don't get broken.
- Improve our read speed from the modbus interface, because we can
  request fewer, larger batches of registers, then walk through them
  formatting them based on which groups of data are actually registers.

I'm not wild about peering inside another library and using a
sort-of-internal-only method, but this makes our code _much_ better.
I'm also not a massive fan of `RegisterSet.read_metrics()` basically
having to maintain a cursor to walk the raw values we get from the
modbus… But it's better than reading them one at a time, and I can't
think of a cleaner way right now.