Reduce contention errors per-request sockets
The problem, it turns out, with using one modbus connection as per
ee36d62 is that if any one request fails and breaks the socket, it
stays broken and is hard to fix. The problem we were trying avoid with
that change was being unable to connect another socket because there
was still a previous connection lingering.

The better solution to that, it turns out, is to tidy up after
ourselves and disconnect the socket when we're done with it (and only
open it when we need it).

This change uses `yield from` to retain the generator-style behaviour
of `get_metrics_reader()`, but returns responsibility for opening (and
closing!) the modbus connection to it. This means that every time
the Prometheus collector endpoint is requested, we open, read from, and
close a fresh socket, meaning every new request can succeed.

The only obvious issue this leaves is the case when a second request
comes in while we're still responding to the first, which can cause one
to fail. Probably the solution to that is to introduce some sort of
locking?