Avoid socket cross-talk by locking
Previously, it was possible (because, I think, of the way the modbus
communication works) to have two parallel requests create two
independent sockets, and each would receive the modbus messages for the
other. This _might_ be solvable by using more robust socket logic, or
by switching to the async version of PySolarmanV5, but I'm not convinced.

For now, we're avoiding the problem entirely by using a lock to control
access to the socket, so only one socket to the data logger can be open
at a time. This means that sometimes a second (or third, and so on)
request will have to wait until the socket lock is free, but given 90%
of the time this will only be hit by one Prometheus collector (unless
I'm testing stuff), that's almost certainly good enough.